refactor: fix internal imports in skills-manager-core
